+++ HISTORY.md
@@ -47,6 +47,7 @@
 
 ### 2022-05-01
 * byob: programmatically reduce the number of inputs in a custom block
+* byob: programmatically add inputs to a custom block
 
 ### 2022-04-28
 * threads, byob: programmatically re-define custom blocks, experimental, under construction